refactor and reorganize files
