+++ projects/schematics/src/modernize-app-migrated-from-68-to-221119/ssr/update-angular-json-for-ssr.ts
@@ -4,8 +4,21 @@ import { getWorkspace } from '../../shared/utils/workspace-utils';
 /**
  * Updates the Angular configuration related to SSR for new Angular v17 standards.
  *
- * Sets up SSR options in the build configuration, adds 'noSsr' mode,
- * and removes obsolete SSR-related architect sections.
+ * 1. In the section `architect > build > options` it adds 3 new options with values:
+ *    `"server": "src/main.server.ts"`,
+ *    `"prerender": false`,
+ *     `"ssr": { "entry": "server.ts" }`
+ *
+ * 2. In the section `architect > build > configurations` it adds a new property with object value
+ *    `"noSsr": { "ssr": false, "prerender": false }`
+ *
+ * 3. In the section `architect > serve > configurations` it adds the ending `,noSsr` (with the preceding comma)
+ *    at the end of the string values in subsections
+ *    `... > production > buildTarget` and
+ *    `... > development > buildTarget`
+ *
+ * 4. It removes the whole 3 sections `architect > server`, `architect > serve-ssr` and `architect > prerender`
+ *    (because their responsibilities are now handled just by the single new Angular `application` builder)
  */
 export function updateAngularJsonForSsr(): Rule {
   return (tree: Tree, context: SchematicContext) => {
@@ -38,7 +51,7 @@ export function updateAngularJsonForSsr(): Rule {
       },
     };
 
-    // Update serve configurations with noSsr
+    // Update serve configurations with `,noSsr`
     const serveConfigs = project.architect?.serve?.configurations;
     if (serveConfigs) {
       project.architect.serve = {
@@ -57,7 +70,7 @@ export function updateAngularJsonForSsr(): Rule {
       } as any;
     }
 
-    // Remove obsolete targets
+    // Remove obsolete architect targets
     delete project.architect?.server;
     delete project.architect?.['serve-ssr'];
     delete project.architect?.prerender;
